original image, html setting

+++ data/htdocs/img_change.php
@@ -5,22 +5,41 @@ $base_dir = dirname(__FILE__);
 
 
 $url_no = $_GET['url_no'];
+$skip = $_GET['skip'];
 
-if(isset($url_no)){
+if(isset($url_no) && $url_no > 0){
 
-    //todo 
-    //image file move
+    #파일 일련번호 형식
     $url_no = sprintf('%04d', $url_no);
+    $move_img_fg = 0;
+    $move_html_fg = 0;
 
+    # 이미지 파일
     $filename = $url_no."_site.png";
     echo "<br>".$daily_file = IMG_CAPTURE_PATH."daily/".$filename;
     echo "<br>".$origin_file = IMG_CAPTURE_PATH."orign/".$filename;
     if(file_exists($daily_file)){
         if(copy($daily_file, $origin_file)){
-            unlink($daily_file);
-            alert_close('적용 완료');
+            #unlink($daily_file);
+            $move_img_fg = 1;
         }	
     }
+
+    # HTML 파일
+    $filename = $url_no.".html";
+    echo "<br>".$daily_file = HTML_CAPTURE_PATH."daily/".$filename;
+    echo "<br>".$origin_file = HTML_CAPTURE_PATH."orign/".$filename;
+    if(file_exists($daily_file)){
+        if(copy($daily_file, $origin_file)){
+            unlink($daily_file); 
+            $move_html_fg = 1;
+        }	
+    }
+
+    if($move_img_fg == 1 && $move_html_fg == 1){
+        alert_close('적용 완료');
+    }
+    
     
 } else {
     alert_close('해당하는 URL이 존재하지 않음');
